+++ src/OpenTelemetry.Api/Context/Propagation/B3Propagator.cs
@@ -25,8 +25,9 @@ namespace OpenTelemetry.Context.Propagation
 {
     /// <summary>
     /// A text map propagator for B3. See https://github.com/openzipkin/b3-propagation.
+    /// This class has been deprecated in favour of OpenTelemetry.Extensions.Propagators package.
     /// </summary>
-    [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
+    [Obsolete("Use B3Propagator class from OpenTelemetry.Extensions.Propagators namespace, shipped as part of the same nuget package.")]
     public sealed class B3Propagator : TextMapPropagator
     {
         internal const string XB3TraceId = "X-B3-TraceId";
@@ -59,7 +60,7 @@ namespace OpenTelemetry.Context.Propagation
         /// <summary>
         /// Initializes a new instance of the <see cref="B3Propagator"/> class.
         /// </summary>
-        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
+        [Obsolete("Use B3Propagator constructor from OpenTelemetry.Extensions.Propagators package instead.")]
         public B3Propagator()
             : this(false)
         {
@@ -69,7 +70,7 @@ namespace OpenTelemetry.Context.Propagation
         /// Initializes a new instance of the <see cref="B3Propagator"/> class.
         /// </summary>
         /// <param name="singleHeader">Determines whether to use single or multiple headers when extracting or injecting span context.</param>
-        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
+        [Obsolete("Use B3Propagator constructor from OpenTelemetry.Extensions.Propagators package instead.")]
         public B3Propagator(bool singleHeader)
         {
             this.singleHeader = singleHeader;
@@ -79,7 +80,7 @@ namespace OpenTelemetry.Context.Propagation
         public override ISet<string> Fields => AllFields;
 
         /// <inheritdoc/>
-        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
+        [Obsolete("Use Extract function from B3Propagator class in OpenTelemetry.Extensions.Propagators package instead.")]
 #pragma warning disable CS0809 // Obsolete member overrides non-obsolete member
         public override PropagationContext Extract<T>(PropagationContext context, T carrier, Func<T, string, IEnumerable<string>> getter)
 #pragma warning restore CS0809 // Obsolete member overrides non-obsolete member
@@ -113,7 +114,7 @@ namespace OpenTelemetry.Context.Propagation
         }
 
         /// <inheritdoc/>
-        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
+        [Obsolete("Use Inject function from B3Propagator class in OpenTelemetry.Extensions.Propagators package instead.")]
 #pragma warning disable CS0809 // Obsolete member overrides non-obsolete member
         public override void Inject<T>(PropagationContext context, T carrier, Action<T, string, string> setter)
 #pragma warning restore CS0809 // Obsolete member overrides non-obsolete member