+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* B3 Propagator class has been deprecated and moved as is to a new
+* `B3Propagator` class has been deprecated and moved as is to a new
  `OpenTelemetry.Extensions.Propagators` package.
  It will be removed in the next major release, see issue [#3259](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3259)